Initial commit of homework

+++ src/com/company/homework1.java
@@ -2,11 +2,11 @@ package com.company;
 
 public class homework1 {
     public static void main (String[] args){
-        double[] arr = {7, 5, 3, 8, 2, 9};
+        int[] arr = {7, 5, 3, 8, 2, 9};
         System.out.println(average(arr));
     }
 
-    static double average(double[] arr) {
+    static double average(int[] arr) {
         double sum = 0;
         int i;
         for (i = 0; i < arr.length; i++) {
